acc-7 introduce HasMarylandIncome
